% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_shp.R
\name{clip_raster}
\alias{clip_raster}
\title{clip raster by mask or points}
\usage{
clip_raster(x, y, return.id = TRUE)
}
\arguments{
\item{x}{Raster* object}

\item{y}{points represented by a two-column matrix or data.frame, or SpatialPoints*;
SpatialPolygons*; SpatialLines; Extent; or a numeric vector representing cell numbers}
}
\description{
clip raster by mask or points
}
