% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extractId.R
\name{df2sp}
\alias{df2sp}
\title{df2sp}
\usage{
df2sp(d, formula = ~lon + lat, prj)
}
\arguments{
\item{d}{A data.frame with coordinates information}

\item{formula}{~longitude+latitude}

\item{prj}{\code{\link[sp:CRS-class]{sp::CRS-class()}} Projection.}
}
\description{
Convert dataframe data into SpatialPointsDataframe (Deprecated! Use \code{\link[=df2sf]{df2sf()}}
instead.)
}
