% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\alias{get_grid.lonlat}
\alias{fix_lon}
\title{get_grid}
\usage{
get_grid(range, cellsize, midgrid = c(TRUE, TRUE), prj = prj84,
  type = "mat", fix_lon360 = FALSE)

get_grid.lonlat(lon, lat, fix_lon360 = FALSE)

fix_lon(x)
}
\arguments{
\item{range}{A numeric vector, \code{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{Numeric vector, grid cell size \code{[cellsize_lon, cellsize_lat]}.}

\item{midgrid}{A vector, \code{[midgrid_lon, midgrid_lat]}. If midgrid = false, then
begin point and end point locate on grid lines; If true, then begin point and
end point in the middle of grid.}

\item{prj}{\code{\link[sp:CRS-class]{sp::CRS-class()}} Projection.}

\item{type}{Character, one of \code{mat} or \code{vec}.
\itemize{
\item \code{mat}: If \code{image(x)} looks correct, x can be put into grid directly:
\code{grid@data <- data.frame(x = as.numeric(x))}
\item \code{vec}: If \code{image(x)} looks correct, x needs to \code{flipud}:
\code{grid@data <- data.frame(x = as.numeric(flipud(x)))}
}}

\item{fix_lon360}{boolean}
}
\description{
construct SpatialGridDataFrame
}
\details{
The input matrix suits for \code{image}, can be directly input to \code{get_grid}.
}
\examples{
range <- c(73, 105, 25, 40) # Tibetan Plateau
grid  <- get_grid(range, cellsize = 1/2, midgrid = TRUE)
}
