% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP6tools.R
\name{get_RegionalMean}
\alias{get_RegionalMean}
\alias{get_RegionalMeans}
\title{get_RegionalMean}
\usage{
get_RegionalMean(f, shp = sf2::bou1_4p)

get_RegionalMeans(fs, shp = sf2::bou1_4p)
}
\description{
get_RegionalMean
}
