% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_MODIS.R
\name{getPointsMODISTile}
\alias{getPointsMODISTile}
\title{Get points data from MODIS tiles}
\usage{
getPointsMODISTile(coor, cellsize = 1/240)
}
\arguments{
\item{coor}{2-column matrix of coordinates (longitude, latitude)}

\item{cellsize}{500m products, cellsize is regarded as 1/240; 1km product,
cellsize is regarded as 1/120.}
}
\value{
\itemize{
\item \code{h} MODIS tile horizontal id
\item \code{v} MODIS tile vertical id
\item \code{i} point horizontal position at the tile
\item \code{j} point vertical position at the tile
}
}
\description{
Get points data from MODIS tiles
}
\examples{
coor <- cbind(101:120, 21:40)
d <- getPointsMODISTile(coor)
}
\references{
\enumerate{
\item https://landweb.modaps.eosdis.nasa.gov/cgi-bin/developer/tilemap.cgi
\item https://modis-land.gsfc.nasa.gov/MODLAND_grid.html
}
}
