% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_basin.R
\name{guess_network}
\alias{guess_network}
\title{guess_network}
\usage{
guess_network(info)
}
\arguments{
\item{info:}{the first two columns should be \verb{[i, contains]}}
}
\description{
guess_network
}
\examples{
\dontrun{
info = shp_contains(shp)
l_net = guess_networ(info)
}
}
\references{
https://stackoverflow.com/questions/25130462/get-disjoint-sets-from-a-list-in-r
}
