% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwST.R
\name{idwST}
\alias{idwST}
\alias{idwST.spatial}
\alias{idwST.formula}
\title{Inverse Distance Weighting (IDW) function for spatio-temporal prediction.}
\usage{
idwST.spatial(location, newdata, data, maxdist = Inf, nmax = NULL,
  factor.p = 2)

idwST.formula(formula, location, newdata, maxdist = Inf, nmax = NULL,
  factor.p = 2)
}
\arguments{
\item{location}{SpatialPointsDataFrame}

\item{newdata}{location frame or spatial object with prediction/simulation
spatio-temporal locations; should contain attribute columns with the
independent variables (if present) and (if locations is a formula) the
coordinates and time with names, as defined in locations where you want to
generate new predictions}

\item{data}{data.frame, \code{[ntime, 1+nsite]}, 1th column is \code{date} object.
should contain the spatio-temporal dependent
variable, independent variables (statics and/or dynamics), spatial coordinates
and the time as an integer or numerical variable.}

\item{maxdist}{max distance (km)}

\item{nmax}{number of nearest observations that should be used for a \emph{idwST}
prediction, where nearest is defined in terms of the spatio-temporal locations}

\item{factor.p}{numeric; specify the inverse distance weighting power (\emph{p}
is the exponent that influences the weighting or optimal smoothing parameter)}

\item{formula}{formula that defines a detrended linear model, use \eqn{z_{st}}\code{~1}.}
}
\value{
Attributes columns contain coordinates, time, predictions, and the variance
column contains NA's.
}
\description{
This function performs spatio-temporal interpolation. Here \emph{idwST} is in a
local neighborhood. This interpolation method considers the value of a point can
be obtained from the weighted sum of values of the regionalized variable of
closest neighbors. The general formula for the IDW is given by: \deqn{
\hat{z}_0(st)=\sum_{i=1}^n \lambda_i z_i(st) } The expression for determining
the weights is: \deqn{ \lambda_i = \frac{d_{i0}^{-p}}{\sum_{i=1}^n d_{i0}^{-p}}
} The weight is controlled by a factor \emph{p} with each increment of the
distance, \eqn{d_{i0}} is the distance between the prediction position and each
of the measured positions.

The expression \eqn{d_{i0}} can be obtained by:
\deqn{d_{i0}=\sqrt{(x_{i}-x_{0})^2+(y_{i}-y_{0})^2+C\cdot (t_{i}-t_{0})^2} } \eqn{x},
\eqn{y} and \eqn{t} correspond to the spatio-temporal coordinates, \emph{p}
(factor.p).
}
\details{
idwST function generates individual spatio-temporal predictions from
IDW spatio-temporal interpolation. IDW is a type of deterministic method for
interpolation, the assigned values to unknown points are calculated with a
weighted average of the values available at the known points.
}
\examples{
\dontrun{
print('hello')
r <- idwST(sp_all, data, gridclip)
}
}
\references{
\enumerate{
\item Li L, Losser T, Yorke C, Piltner R. (2014). \emph{Fast inverse distance
weighting-based spatiotemporal interpolation: a web-based application of
interpolating daily fine particulate matter PM2:5 in the contiguous U.S.
using parallel programming and k-d tree.} Int. J. Environ. Res. Public Health,
11: 9101-9141. \href{http://www.mdpi.com/journal/ijerph}{link}
}
}
