% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwST0.R
\name{idwST0}
\alias{idwST0}
\title{Inverse Distance Weighting (IDW)}
\usage{
idwST0(locations, newdata, mat_z, maxdist = Inf, nmax = NULL, factor.p)
}
\description{
Inverse Distance Weighting (IDW)
}
\examples{
\dontrun{
r <- idwST0(location, newdata, mat_z, maxdist, nmax, factor.p)
}
}
\keyword{internal}
