% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{load_data}
\alias{load_data}
\alias{load_china_river}
\alias{load_china}
\alias{load_continent}
\title{load_data}
\usage{
load_china_river()

load_china()

load_continent()
}
\description{
load_data
}
\details{
\itemize{
\item \link{load_china_river}
}
\itemize{
\item \verb{river$i$}: river polylines
\item \verb{river_p$i$}: river polygons
}
\itemize{
\item 
\item 
}
}
