% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{make_grid for global basins}
\usage{
make_grid(range = c(-180, 180, -60, 90), cellsize, midgrid = c(TRUE, TRUE))
}
\arguments{
\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{Numeric vector, grid cell size \verb{[cellsize_lon, cellsize_lat]}.}

\item{midgrid}{A vector, \verb{[midgrid_lon, midgrid_lat]}. If midgrid = false, then
begin point and end point locate on grid lines; If true, then begin point and
end point in the middle of grid.}
}
\description{
make_grid for global basins
}
\examples{
\dontrun{
r <- make_grid(range=c(-180, 180, -60, 90), cellsize=1)
}
}
