% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rast.R
\name{make_rast}
\alias{make_rast}
\title{make_rast}
\usage{
make_rast(range = c(-180, 180, -90, 90), cellsize = 1, nlyrs = 1, ...)
}
\arguments{
\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{Numeric vector, grid cell size \verb{[cellsize_lon, cellsize_lat]}.}

\item{nlyrs}{positive integer. Number of layers}

\item{...}{other parameters to \code{\link[terra:rast]{terra::rast()}}, e.g., names, vals.}
}
\description{
make_rast
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}
}
