% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwST.R
\name{nanMatMult}
\alias{nanMatMult}
\title{Matrix product with NA values}
\usage{
nanMatMult(x, w, FUN = crossprod)
}
\arguments{
\item{x}{matrix}

\item{w}{matrix, weight.}
}
\description{
\code{matmult_eigenMap} is about 11 times faster than \code{crossprod}.
}
\examples{
\dontrun{
A <- matrix(rnorm(10000), 1000, 1000)
B <- matrix(rnorm(10000), 1000, 1000)

rbenchmark::benchmark(
  r1 <- matmult(A, B, crossprod),
  # r2 <- matmult(A, B, Rfast::mat.mult),
  r3 <- matmult(A, B, matmult_arma),
  r4 <- matmult(A, B, matmult_eigen),
  r5 <- matmult(A, B, matmult_eigenMap),
  replications = 10)
}
}
