% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_fraction.R
\name{overlap_fraction}
\alias{overlap_fraction}
\title{overlap_fraction}
\usage{
overlap_fraction(shp, cellsize = 0.1, range = NULL, outfile = NULL, ...)
}
\value{
\itemize{
\item \code{I}   : the order of \code{base}, \verb{[lon, lat]}, \code{image} works directly on it.
\item \code{cell}: the order of \code{gdal}, \verb{[lon, rev(lat)]}, returned by \code{values(r)},
}
}
\description{
overlap_fraction
}
\examples{
\dontrun{
library(nctools)

f = "//kong-nas/CMIP6/DATA/China/CN0.5.1_ChinaDaily_025x025/yearly/CN05.1_Tmax_1961_2021_yearly_025x025.nc"

data = ncread(f, "tmax")
image(data[, , 1]) # should works

r = rast(f)
range = as.vector(ext(r))
cellsize = res(r)
}

# shp = basin_Baihe
shp = bou1_4p
cellsize <- 0.5
range <- st_range(shp, cellsize)

dat = overlap_fraction(shp, cellsize = cellsize, range, outfile = NULL)
print(dat)
}
