% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPolygonGrids.R
\name{overlapPolygonGrids}
\alias{overlapPolygonGrids}
\title{overlapPolygonGrids}
\usage{
overlapPolygonGrids(shpfile, cellsize = 1/24, show = FALSE)
}
\arguments{
\item{shpfile}{file path of polygon shapefile}

\item{cellsize}{double}

\item{show}{Boolean, whether to show cliped grids?}
}
\value{
\itemize{
\item \code{lon}: longitude of grids
\item \code{lat}: latitude of grids
\item \code{id}: id of grids
}
}
\description{
overlapPolygonGrids
}
\examples{
\dontrun{
d <- overlapBasinGrids(shpfile, show = TRUE)
}
}
