% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_rast.R
\name{rast_df}
\alias{rast_df}
\title{rast_coord}
\usage{
rast_df(r, .area = TRUE, na.rm = FALSE)
}
\value{
\itemize{
\item \code{I}   : the order of \code{base}, \verb{[lon, lat]}, \code{image} works directly on it.
\item \code{cell}: the order of \code{gdal}, \verb{[lon, rev(lat)]}, returned by \code{values(r)},
}
}
\description{
rast_coord
}
