% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shp.R
\name{read_shp}
\alias{read_shp}
\title{read shapefile}
\usage{
read_shp(file, ...)
}
\arguments{
\item{file}{The path of ESRI shapefile}

\item{...}{other parameters to \code{\link[rgdal:readOGR]{rgdal::readOGR()}}}
}
\description{
read shapefile
}
\seealso{
\code{\link[rgdal:readOGR]{rgdal::readOGR()}}
}
