% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2-package.R
\docType{package}
\name{sp2-package}
\alias{sp2}
\alias{sp2-package}
\title{sp2: Spatial Analysis tools}
\description{
sp2 extends the function of sp package.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
