% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dissolve.R
\name{st_dissolve}
\alias{st_dissolve}
\alias{st_dissolve.sf}
\alias{st_dissolve.character}
\title{st_dissolve}
\usage{
st_dissolve(x, by = NULL, ...)

\method{st_dissolve}{sf}(x, by = NULL, ...)

\method{st_dissolve}{character}(x, by = NULL, ...)
}
\description{
st_dissolve
}
