% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extractId.R
\name{st_extractId}
\alias{st_extractId}
\title{st_extractId}
\usage{
st_extractId(x, y, plot = TRUE, ...)
}
\arguments{
\item{x}{A sf point, or file path}

\item{y}{A sf polygon, or file path}

\item{plot}{Boolean. Whether to visualize extracted points?}

\item{...}{others passed to \code{\link[sf:geos_binary_pred]{sf::st_within()}}}
}
\description{
st_extractId
}
