% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_range.R
\name{st_range}
\alias{st_range}
\title{get range of \code{sf} object}
\usage{
st_range(x, cellsize = 0.1)
}
\arguments{
\item{x}{\code{sf} object}

\item{cellsize}{double}
}
\description{
get range of \code{sf} object
}
