% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rect.R
\name{st_rect}
\alias{st_rect}
\title{st_rect}
\usage{
st_rect(range, crs = st_crs(4326))
}
\arguments{
\item{range}{\verb{[lon_min, lon_max, lat_min, lat_max]}}
}
\description{
st_rect
}
\examples{
\dontrun{
xlim = c(112, 115)
ylim = c(21, 23)
range <- c(xlim, ylim)
poly = st_rect(range)
write_sf(poly, "poly.shp")
}
}
