% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_360.R, R/make_grid.R
\name{fix_lon}
\alias{fix_lon}
\alias{make_grid}
\alias{get_grid.lonlat}
\title{make_grid}
\usage{
fix_lon(x)

make_grid(
  range = c(-180, 180, -90, 90),
  cellsize = 1,
  midgrid = c(TRUE, TRUE),
  prj = prj84,
  type = "base",
  fix_lon360 = FALSE
)

get_grid.lonlat(lon, lat, fix_lon360 = FALSE)
}
\arguments{
\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{Numeric vector, grid cell size \verb{[cellsize_lon, cellsize_lat]}.}

\item{midgrid}{A vector, \verb{[midgrid_lon, midgrid_lat]}. If midgrid = false, then
begin point and end point locate on grid lines; If true, then begin point and
end point in the middle of grid.}

\item{prj}{\code{\link[sp:CRS-class]{sp::CRS-class()}} Projection.}

\item{type}{Character, one of \code{mat} or \code{vec}.
\itemize{
\item \code{base}: If \code{image(x)} looks correct, x can be put into grid directly:
\code{grid@data <- data.frame(x = as.numeric(x))}
\item \code{gdal}: If \code{image(x)} looks correct, x needs to \code{flipud}:
\code{grid@data <- data.frame(x = as.numeric(Ipaper::flipud(x)))}
In version >= 0.1.3, \code{c("mat", "vec")} was renamed as \code{c("base", "gdal")}.
}}

\item{fix_lon360}{boolean}
}
\description{
construct SpatialGridDataFrame
}
\details{
The input matrix suits for \code{image}, can be directly input to \code{make_grid}.
}
\examples{
range <- c(73, 105, 25, 40) # Tibetan Plateau
grid  <- make_grid(range, cellsize = 1/2, midgrid = TRUE)
}
