% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPolygonGrids.R
\name{overlapPolygonGrids}
\alias{overlapPolygonGrids}
\alias{overlapPolygonGrids2}
\title{overlapPolygonGrids}
\usage{
overlapPolygonGrids(shp, cellsize = 1/24, type = "big", show = FALSE)

overlapPolygonGrids2(shp, cellsize = 1/24, show = FALSE)
}
\arguments{
\item{cellsize}{double}

\item{show}{Boolean, whether to show cliped grids?}

\item{shpfile}{file path of polygon shapefile}
}
\value{
\itemize{
\item \code{lon}: longitude of grids
\item \code{lat}: latitude of grids
\item \code{id}: id of grids
}
}
\description{
overlapPolygonGrids
}
\examples{
\dontrun{
shpfile <- "D:/Documents/Tencent Files/991810576/FileRecv/jianyu/shp/af_0000001.shp"
shp <- rgdal::readOGR(shpfile, verbose = FALSE)
d <- overlapPolygonGrids(shp, show = TRUE)
}
}
