% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_rast.R
\name{rast_df}
\alias{rast_df}
\title{rast_coord}
\usage{
rast_df(r, .area = TRUE, na.rm = FALSE)
}
\description{
rast_coord
}
\details{
\itemize{
\item \code{I}: the order of \code{values(r)}
\item \code{cell}: the order of \code{values(r)}
}
}
