% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{st_read}
\alias{st_write}
\alias{res}
\alias{rast}
\alias{plot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sf}{\code{\link[sf]{st_read}}, \code{\link[sf]{st_write}}}

  \item{terra}{\code{\link[terra]{plot}}, \code{\link[terra]{rast}}, \code{\link[terra:dimensions]{res}}}
}}

