% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shp.R
\name{write_shp}
\alias{write_shp}
\alias{write_shp_sp}
\title{write and read ESRI Shapefile}
\usage{
write_shp(x, file, ...)

write_shp_sp(x, file, ...)
}
\arguments{
\item{x}{Spatial* Object from sp package}

\item{file}{The path of ESRI shapefile}

\item{...}{other parameters to \code{\link[rgdal:rgdal-deprecated]{rgdal::writeOGR()}}}
}
\description{
write and read ESRI Shapefile
}
